refactor(torghut): split profitability frontier ownership
