+++ services/torghut/scripts/consistent_profitability_frontier/__init__.py
@@ -1,0 +1,1 @@
+"""Explicit owner modules for consistent profitability frontier search."""
