+++ services/torghut/scripts/search_consistent_profitability_frontier.py
@@ -78,8 +78,63 @@ from scripts.search_profitability_frontier import (
     apply_candidate_to_configmap,
 )
 
+from scripts.consistent_profitability_frontier.common import (
+    _SECOND_OOS_WINDOW_ID as _SECOND_OOS_WINDOW_ID,
+    FullWindowConsistencyPolicy as FullWindowConsistencyPolicy,
+    OrderTypeAblationPolicy as OrderTypeAblationPolicy,
+    _write_json_output as _write_json_output,
+    _replay_tape_selection_metadata as _replay_tape_selection_metadata,
+    _resolve_full_window as _resolve_full_window,
+    _max_drawdown_from_daily_net as _max_drawdown_from_daily_net,
+    _daily_filled_notional as _daily_filled_notional,
+    _daily_liquidity_notional as _daily_liquidity_notional,
+    _daily_decimal_metric as _daily_decimal_metric,
+    _daily_int_metric as _daily_int_metric,
+    _int_mapping as _int_mapping,
+    _mapping as _mapping,
+    _optional_decimal as _optional_decimal,
+    _nonnegative_int_metric as _nonnegative_int_metric,
+    _truthy_metric as _truthy_metric,
+)
+from scripts.consistent_profitability_frontier.ledger_order import (
+    _order_lifecycle_metrics as _order_lifecycle_metrics,
+    _order_type_execution_metrics as _order_type_execution_metrics,
+    _normalized_order_type as _normalized_order_type,
+    _selected_entry_order_type as _selected_entry_order_type,
+    _forced_order_type_sample_count as _forced_order_type_sample_count,
+    _payload_digest as _payload_digest,
+    _artifact_run_dir_name as _artifact_run_dir_name,
+    _order_type_ablation_artifact_dir as _order_type_ablation_artifact_dir,
+    _frontier_ledger_text as _frontier_ledger_text,
+    _frontier_ledger_datetime as _frontier_ledger_datetime,
+    _frontier_exact_replay_bucket_range as _frontier_exact_replay_bucket_range,
+    _frontier_exact_replay_rows as _frontier_exact_replay_rows,
+    _frontier_exact_replay_bucket_has_authority as _frontier_exact_replay_bucket_has_authority,
+    _frontier_exact_replay_bucket as _frontier_exact_replay_bucket,
+    _exact_replay_ledger_artifact_update as _exact_replay_ledger_artifact_update,
+    _order_type_replay_arm_summary as _order_type_replay_arm_summary,
+    _order_type_ablation_payload as _order_type_ablation_payload,
+)
+from scripts.consistent_profitability_frontier.stress_metrics import (
+    DELAY_ADJUSTED_DEPTH_STRESS_GRID_MS as DELAY_ADJUSTED_DEPTH_STRESS_GRID_MS,
+    CONFORMAL_TAIL_RISK_ALPHA as CONFORMAL_TAIL_RISK_ALPHA,
+    BREAKEVEN_TRANSACTION_COST_BUFFER_MIN_BPS as BREAKEVEN_TRANSACTION_COST_BUFFER_MIN_BPS,
+    MARKET_IMPACT_STRESS_SOURCE_MARKERS as MARKET_IMPACT_STRESS_SOURCE_MARKERS,
+    _p10 as _p10,
+    _conformal_tail_loss_buffer as _conformal_tail_loss_buffer,
+    _conformal_tail_risk_metrics as _conformal_tail_risk_metrics,
+    _breakeven_transaction_cost_buffer_metrics as _breakeven_transaction_cost_buffer_metrics,
+    _delay_depth_fillability as _delay_depth_fillability,
+    _implementation_uncertainty_metrics as _implementation_uncertainty_metrics,
+    _replay_stress_metrics as _replay_stress_metrics,
+    _decimal_payload_metric as _decimal_payload_metric,
+    _max_best_day_share_of_total_pnl as _max_best_day_share_of_total_pnl,
+    _consistency_penalty as _consistency_penalty,
+    _second_oos_summary as _second_oos_summary,
+    _holdout_oos_passed as _holdout_oos_passed,
+)
+
 _LOCAL_ONLY_OVERRIDE_KEYS = frozenset({"normalization_regime"})
-_SECOND_OOS_WINDOW_ID = "second_oos"
 _SAFE_EXACT_REPLAY_CANDIDATE_CAP = 6
 _SAFE_EXACT_REPLAY_EXPLOITATION_SLOTS = 4
 _SAFE_EXACT_REPLAY_EXPLORATION_SLOTS = 2
@@ -265,74 +320,6 @@ class _WorklistItem:
 
 
 @dataclass(frozen=True)
-class FullWindowConsistencyPolicy:
-    target_net_per_day: Decimal
-    min_daily_net_pnl: Decimal
-    min_active_days: int
-    min_active_ratio: Decimal
-    min_positive_days: int
-    max_worst_day_loss: Decimal
-    max_negative_days: int
-    max_drawdown: Decimal
-    max_best_day_share_of_total_pnl: Decimal
-    min_avg_filled_notional_per_day: Decimal
-    min_avg_filled_notional_per_active_day: Decimal
-    require_every_day_active: bool
-    min_regime_slice_pass_rate: Decimal = Decimal("0")
-    max_symbol_concentration_share: Decimal = Decimal("1")
-    max_entry_family_contribution_share: Decimal = Decimal("1")
-    max_gross_exposure_pct_equity: Decimal = Decimal("999999999")
-    min_cash: Decimal = Decimal("-999999999")
-    min_window_weekday_count: int = 0
-
-    def to_payload(self) -> dict[str, Any]:
-        return {
-            "target_net_per_day": str(self.target_net_per_day),
-            "min_daily_net_pnl": str(self.min_daily_net_pnl),
-            "min_active_days": self.min_active_days,
-            "min_active_ratio": str(self.min_active_ratio),
-            "min_positive_days": self.min_positive_days,
-            "max_worst_day_loss": str(self.max_worst_day_loss),
-            "max_negative_days": self.max_negative_days,
-            "max_drawdown": str(self.max_drawdown),
-            "max_best_day_share_of_total_pnl": str(
-                self.max_best_day_share_of_total_pnl
-            ),
-            "min_avg_filled_notional_per_day": str(
-                self.min_avg_filled_notional_per_day
-            ),
-            "min_avg_filled_notional_per_active_day": str(
-                self.min_avg_filled_notional_per_active_day
-            ),
-            "require_every_day_active": self.require_every_day_active,
-            "min_regime_slice_pass_rate": str(self.min_regime_slice_pass_rate),
-            "max_symbol_concentration_share": str(self.max_symbol_concentration_share),
-            "max_entry_family_contribution_share": str(
-                self.max_entry_family_contribution_share
-            ),
-            "max_gross_exposure_pct_equity": str(self.max_gross_exposure_pct_equity),
-            "min_cash": str(self.min_cash),
-            "min_window_weekday_count": self.min_window_weekday_count,
-        }
-
-
-@dataclass(frozen=True)
-class OrderTypeAblationPolicy:
-    enabled: bool
-    max_candidates: int
-    min_sample_count: int
-    max_opportunity_cost_bps: Decimal
-
-    def to_payload(self) -> dict[str, Any]:
-        return {
-            "enabled": self.enabled,
-            "max_candidates": self.max_candidates,
-            "min_sample_count": self.min_sample_count,
-            "max_opportunity_cost_bps": str(self.max_opportunity_cost_bps),
-        }
-
-
-@dataclass(frozen=True)
 class FrontierReplayWindows:
     train_days: tuple[date, ...]
     holdout_days: tuple[date, ...]
@@ -395,11 +382,6 @@ def _order_type_ablation_policy(
             str(value.get("max_opportunity_cost_bps", "8"))
         ),
     )
-
-
-def _write_json_output(path: Path, payload: Mapping[str, Any]) -> None:
-    path.parent.mkdir(parents=True, exist_ok=True)
-    path.write_text(json.dumps(payload, indent=2, sort_keys=True), encoding="utf-8")
 
 
 def _safe_exact_replay_candidate_budget(raw_value: Any) -> int:
@@ -1453,37 +1435,6 @@ def _candidate_evaluation_key_payload(
     return payload
 
 
-def _replay_tape_selection_metadata(
-    validation: Mapping[str, Any] | None,
-) -> dict[str, Any]:
-    payload = dict(validation or {})
-    return {
-        "content_sha256": str(payload.get("content_sha256") or ""),
-        "dataset_snapshot_ref": str(payload.get("dataset_snapshot_ref") or ""),
-        "source_query_digest": str(payload.get("source_query_digest") or ""),
-        "source_table_versions": dict(
-            cast(Mapping[str, Any], payload.get("source_table_versions") or {})
-        ),
-        "feature_schema_hash": str(payload.get("feature_schema_hash") or ""),
-        "cost_model_hash": str(payload.get("cost_model_hash") or ""),
-        "strategy_family": str(payload.get("strategy_family") or ""),
-        "feature_versions": dict(
-            cast(Mapping[str, Any], payload.get("feature_versions") or {})
-        ),
-        "replay_cache_key": str(payload.get("replay_cache_key") or ""),
-        "cache_identity": dict(
-            cast(Mapping[str, Any], payload.get("cache_identity") or {})
-        ),
-        "selected_symbols": list(
-            cast(Sequence[Any], payload.get("selected_symbols") or ())
-        ),
-        "selected_row_count": int(payload.get("selected_row_count") or 0),
-        "validation_status": str(payload.get("status") or ""),
-        "manifest_start_date": str(payload.get("manifest_start_date") or ""),
-        "manifest_end_date": str(payload.get("manifest_end_date") or ""),
-    }
-
-
 def _resolve_prefetch_symbols(
     *,
     cli_symbols: tuple[str, ...],
@@ -1811,1622 +1762,6 @@ def apply_candidate_to_configmap_with_overrides(
 
     data["strategies.yaml"] = yaml.safe_dump(catalog, sort_keys=False)
     return root
-
-
-def _resolve_full_window(
-    *,
-    args: argparse.Namespace,
-    train_days: tuple[date, ...],
-    holdout_days: tuple[date, ...],
-) -> tuple[date, date]:
-    if str(args.full_window_start_date or "").strip():
-        start = date.fromisoformat(str(args.full_window_start_date))
-    else:
-        start = train_days[0]
-    if str(args.full_window_end_date or "").strip():
-        end = date.fromisoformat(str(args.full_window_end_date))
-    else:
-        end = holdout_days[-1]
-    if start > end:
-        raise ValueError("full_window_invalid_range")
-    return (start, end)
-
-
-def _max_drawdown_from_daily_net(daily_net: Mapping[str, Decimal]) -> Decimal:
-    equity = Decimal("0")
-    peak = Decimal("0")
-    max_drawdown = Decimal("0")
-    for trading_day in sorted(daily_net):
-        equity += daily_net[trading_day]
-        if equity > peak:
-            peak = equity
-        drawdown = peak - equity
-        if drawdown > max_drawdown:
-            max_drawdown = drawdown
-    return max_drawdown
-
-
-def _daily_filled_notional(payload: Mapping[str, Any]) -> dict[str, Decimal]:
-    daily_payload = cast(Mapping[str, Any], payload.get("daily") or {})
-    filled_notional: dict[str, Decimal] = {}
-    for day, value in daily_payload.items():
-        if not isinstance(value, Mapping):
-            continue
-        value_mapping = cast(Mapping[str, Any], value)
-        filled_notional[str(day)] = Decimal(
-            str(value_mapping.get("filled_notional", "0"))
-        )
-    return filled_notional
-
-
-def _daily_liquidity_notional(payload: Mapping[str, Any]) -> dict[str, Decimal]:
-    daily_payload = cast(Mapping[str, Any], payload.get("daily") or {})
-    liquidity_notional: dict[str, Decimal] = {}
-    for day, value in daily_payload.items():
-        if not isinstance(value, Mapping):
-            continue
-        value_mapping = cast(Mapping[str, Any], value)
-        raw_value = (
-            value_mapping.get("adv_notional")
-            or value_mapping.get("daily_adv_notional")
-            or value_mapping.get("depth_notional")
-            or value_mapping.get("fillable_depth_notional")
-        )
-        if raw_value is None:
-            continue
-        liquidity_notional[str(day)] = Decimal(str(raw_value))
-    return liquidity_notional
-
-
-def _daily_decimal_metric(payload: Mapping[str, Any], key: str) -> dict[str, Decimal]:
-    daily_payload = cast(Mapping[str, Any], payload.get("daily") or {})
-    values: dict[str, Decimal] = {}
-    for day, value in daily_payload.items():
-        if not isinstance(value, Mapping):
-            continue
-        raw_value = cast(Mapping[str, Any], value).get(key)
-        if raw_value is None:
-            continue
-        values[str(day)] = Decimal(str(raw_value))
-    return values
-
-
-def _daily_int_metric(payload: Mapping[str, Any], key: str) -> dict[str, int]:
-    daily_payload = cast(Mapping[str, Any], payload.get("daily") or {})
-    values: dict[str, int] = {}
-    for day, value in daily_payload.items():
-        if not isinstance(value, Mapping):
-            continue
-        raw_value = cast(Mapping[str, Any], value).get(key)
-        if raw_value is None:
-            continue
-        values[str(day)] = int(raw_value)
-    return values
-
-
-def _int_mapping(value: Any) -> dict[str, int]:
-    if not isinstance(value, Mapping):
-        return {}
-    counts: dict[str, int] = {}
-    for key, item in cast(Mapping[Any, Any], value).items():
-        try:
-            count = int(float(str(item or 0)))
-        except (TypeError, ValueError):
-            count = 0
-        normalized_key = str(key or "").strip().lower()
-        if normalized_key:
-            counts[normalized_key] = count
-    return counts
-
-
-def _mapping(value: Any) -> dict[str, Any]:
-    if not isinstance(value, Mapping):
-        return {}
-    return {str(key): item for key, item in cast(Mapping[Any, Any], value).items()}
-
-
-def _optional_decimal(value: Any) -> Decimal | None:
-    if value in (None, ""):
-        return None
-    return Decimal(str(value))
-
-
-def _nonnegative_int_metric(value: Any) -> int:
-    if value in (None, ""):
-        return 0
-    try:
-        return max(0, int(Decimal(str(value))))
-    except (InvalidOperation, ValueError):
-        return 0
-
-
-def _truthy_metric(value: Any) -> bool:
-    if isinstance(value, bool):
-        return value
-    if value in (None, ""):
-        return False
-    return str(value).strip().lower() in {"1", "true", "yes", "y", "pass", "passed"}
-
-
-def _order_lifecycle_metrics(payload: Mapping[str, Any]) -> dict[str, Any]:
-    lifecycle = _mapping(payload.get("order_lifecycle"))
-    if not lifecycle:
-        return {}
-    queue_ahead_depletion_sample_count = _nonnegative_int_metric(
-        lifecycle.get("queue_ahead_depletion_sample_count")
-        or lifecycle.get("queue_depletion_sample_count")
-    )
-    queue_ahead_depletion_evidence_present = _truthy_metric(
-        lifecycle.get("queue_ahead_depletion_evidence_present")
-    ) or (
-        queue_ahead_depletion_sample_count > 0
-        and (
-            lifecycle.get("queue_ahead_depletion_rate") is not None
-            or lifecycle.get("queue_ahead_depleted_qty_p50") is not None
-            or lifecycle.get("queue_ahead_depletion_time_ms_p50") is not None
-        )
-    )
-    metrics: dict[str, Any] = {
-        "order_lifecycle": lifecycle,
-        "fill_survival_evidence_present": bool(
-            lifecycle.get("fill_survival_evidence_present")
-        ),
-        "fill_survival_sample_count": int(
-            lifecycle.get("fill_survival_sample_count")
-            or lifecycle.get("submitted_order_count")
-            or 0
-        ),
-        "fill_survival_fill_rate": str(lifecycle.get("fill_rate") or "0"),
-        "fill_time_ms_avg": str(lifecycle.get("fill_time_ms_avg") or ""),
-        "fill_time_ms_p50": lifecycle.get("fill_time_ms_p50"),
-        "fill_time_ms_p95": lifecycle.get("fill_time_ms_p95"),
-        "pending_age_ms_p95": lifecycle.get("pending_age_ms_p95"),
-        "max_censored_pending_age_ms": lifecycle.get("max_censored_pending_age_ms"),
-        "spread_bps_avg_at_order": str(lifecycle.get("spread_bps_avg_at_order") or ""),
-        "spread_bps_p95_at_order": str(lifecycle.get("spread_bps_p95_at_order") or ""),
-        "depth_notional_min_at_order": str(
-            lifecycle.get("depth_notional_min_at_order") or ""
-        ),
-        "depth_notional_avg_at_order": str(
-            lifecycle.get("depth_notional_avg_at_order") or ""
-        ),
-        "queue_touch_qty_avg": str(lifecycle.get("queue_touch_qty_avg") or ""),
-        "queue_touch_notional_avg": str(
-            lifecycle.get("queue_touch_notional_avg") or ""
-        ),
-        "order_qty_to_touch_qty_ratio_p95": str(
-            lifecycle.get("order_qty_to_touch_qty_ratio_p95") or ""
-        ),
-        "queue_ahead_depletion_evidence_present": (
-            queue_ahead_depletion_evidence_present
-        ),
-        "queue_ahead_depletion_sample_count": queue_ahead_depletion_sample_count,
-        "delay_adjusted_depth_queue_ahead_depletion_evidence_present": (
-            queue_ahead_depletion_evidence_present
-        ),
-        "delay_adjusted_depth_queue_ahead_depletion_sample_count": (
-            queue_ahead_depletion_sample_count
-        ),
-        "fill_probability_by_latency_bucket": _mapping(
-            lifecycle.get("fill_probability_by_latency_bucket")
-        ),
-        "fill_probability_by_latency_threshold_ms": _mapping(
-            lifecycle.get("fill_probability_by_latency_threshold_ms")
-        ),
-    }
-    for key in (
-        "queue_ahead_depletion_rate",
-        "queue_ahead_qty_p95",
-        "queue_ahead_depleted_qty_p50",
-        "queue_ahead_depleted_qty_p95",
-        "queue_ahead_depletion_time_ms_p50",
-        "queue_ahead_depletion_time_ms_p95",
-    ):
-        if key in lifecycle:
-            metrics[key] = lifecycle[key]
-    survivorship = _mapping(lifecycle.get("post_cost_survivorship"))
-    if survivorship:
-        metrics["post_cost_survivorship"] = survivorship
-        metrics["post_cost_survival_rate"] = str(
-            survivorship.get("post_cost_survival_rate") or "0"
-        )
-        metrics["gross_positive_killed_by_cost_count"] = int(
-            survivorship.get("gross_positive_killed_by_cost_count") or 0
-        )
-    return metrics
-
-
-def _order_type_execution_metrics(payload: Mapping[str, Any]) -> dict[str, Any]:
-    decision_counts = _int_mapping(payload.get("decision_count_by_order_type"))
-    filled_counts = _int_mapping(payload.get("filled_count_by_order_type"))
-    market_decision_count = max(0, decision_counts.get("market", 0))
-    limit_decision_count = max(0, decision_counts.get("limit", 0))
-    market_limit_sample_count = market_decision_count + limit_decision_count
-    metrics: dict[str, Any] = {}
-    if decision_counts:
-        metrics["decision_count_by_order_type"] = decision_counts
-    if filled_counts:
-        metrics["filled_count_by_order_type"] = filled_counts
-    if "limit_fill_rate" in payload:
-        metrics["limit_fill_rate"] = str(payload.get("limit_fill_rate") or "0")
-    if market_limit_sample_count > 0:
-        metrics["market_limit_order_mix_sample_count"] = market_limit_sample_count
-        metrics["market_limit_order_mix_evidence_present"] = True
-    if market_decision_count > 0 and limit_decision_count > 0:
-        metrics["market_limit_order_mix_passed"] = True
-    if limit_decision_count > 0:
-        metrics["limit_fill_probability_sample_count"] = limit_decision_count
-        metrics["limit_fill_probability_evidence_present"] = (
-            "limit_fill_rate" in payload or filled_counts.get("limit", 0) > 0
-        )
-    return metrics
-
-
-def _normalized_order_type(value: Any) -> str:
-    raw_value = str(value or "").strip().lower()
-    if raw_value in {"limit", "prefer_limit"}:
-        return "limit"
-    return "market"
-
-
-def _selected_entry_order_type(
-    *,
-    candidate_params: Mapping[str, Any],
-    strategy_overrides: Mapping[str, Any],
-) -> str:
-    if "entry_order_type" in candidate_params:
-        return _normalized_order_type(candidate_params.get("entry_order_type"))
-    if "entry_order_type" in strategy_overrides:
-        return _normalized_order_type(strategy_overrides.get("entry_order_type"))
-    return "market"
-
-
-def _forced_order_type_sample_count(
-    payload: Mapping[str, Any],
-    *,
-    order_type: str,
-) -> int:
-    decision_counts = _int_mapping(payload.get("decision_count_by_order_type"))
-    if decision_counts:
-        return max(0, decision_counts.get(order_type, 0))
-    return max(0, int(payload.get("decision_count") or 0))
-
-
-def _payload_digest(payload: Mapping[str, Any]) -> str:
-    return hashlib.sha256(
-        json.dumps(payload, sort_keys=True, default=str).encode("utf-8")
-    ).hexdigest()
-
-
-def _artifact_run_dir_name(json_output: Path) -> str:
-    raw_name = json_output.stem.strip()
-    safe_name = "".join(
-        character if character.isalnum() or character in "._-" else "-"
-        for character in raw_name
-    ).strip(".-_")
-    return safe_name or "frontier-run"
-
-
-def _order_type_ablation_artifact_dir(
-    *,
-    args: argparse.Namespace,
-    root: Path,
-) -> Path:
-    json_output = getattr(args, "json_output", None)
-    if isinstance(json_output, Path):
-        return (
-            json_output.parent
-            / "frontier-artifacts"
-            / _artifact_run_dir_name(json_output)
-        )
-    return root / "frontier-artifacts"
-
-
-def _frontier_ledger_text(value: Any) -> str:
-    return str(value or "").strip()
-
-
-def _frontier_ledger_datetime(value: Any, *, date_end: bool = False) -> datetime | None:
-    text = _frontier_ledger_text(value)
-    if not text:
-        return None
-    try:
-        if "T" not in text and len(text) == 10:
-            parsed_date = date.fromisoformat(text)
-            parsed = datetime.combine(parsed_date, time.min, tzinfo=timezone.utc)
-            return parsed + timedelta(days=1) if date_end else parsed
-        parsed = datetime.fromisoformat(text.replace("Z", "+00:00"))
-    except ValueError:
-        return None
-    if parsed.tzinfo is None:
-        return parsed.replace(tzinfo=timezone.utc)
-    return parsed.astimezone(timezone.utc)
-
-
-def _frontier_exact_replay_bucket_range(
-    *,
-    ledger_payload: Mapping[str, Any],
-    full_window_start: date | None,
-    full_window_end: date | None,
-) -> tuple[datetime, datetime] | None:
-    full_window = _mapping(ledger_payload.get("full_window"))
-    start = _frontier_ledger_datetime(
-        ledger_payload.get("window_start")
-        or ledger_payload.get("bucket_started_at")
-        or ledger_payload.get("started_at")
-        or ledger_payload.get("start")
-        or full_window.get("start_date")
-        or (full_window_start.isoformat() if full_window_start is not None else "")
-    )
-    end = _frontier_ledger_datetime(
-        ledger_payload.get("window_end")
-        or ledger_payload.get("bucket_ended_at")
-        or ledger_payload.get("ended_at")
-        or ledger_payload.get("end")
-        or full_window.get("end_date")
-        or (full_window_end.isoformat() if full_window_end is not None else ""),
-        date_end=True,
-    )
-    if start is None or end is None or end <= start:
-        return None
-    return (start, end)
-
-
-def _frontier_exact_replay_rows(
-    ledger_payload: Mapping[str, Any],
-    raw_rows: Sequence[object],
-) -> list[Mapping[str, object]]:
-    defaults = {
-        key: ledger_payload.get(key)
-        for key in (
-            "account_label",
-            "source",
-            "execution_policy_hash",
-            "cost_model_hash",
-            "lineage_hash",
-            "replay_data_hash",
-            "cost_basis",
-        )
-        if ledger_payload.get(key) not in (None, "")
-    }
-    rows: list[Mapping[str, object]] = []
-    for raw_row in raw_rows:
-        if not isinstance(raw_row, Mapping):
-            return []
-        row = dict(cast(Mapping[str, object], raw_row))
-        for key, value in defaults.items():
-            row.setdefault(key, value)
-        rows.append(row)
-    return rows
-
-
-def _frontier_exact_replay_bucket_has_authority(bucket: RuntimeLedgerBucket) -> bool:
-    return (
-        not bucket.blockers
-        and bucket.fill_count > 0
-        and bucket.decision_count > 0
-        and bucket.submitted_order_count > 0
-        and bucket.closed_trade_count > 0
-        and bucket.open_position_count == 0
-        and bucket.filled_notional > 0
-        and bucket.post_cost_expectancy_bps is not None
-        and bool(bucket.cost_basis_counts)
-        and bool(bucket.execution_policy_hash_counts)
-        and bool(bucket.cost_model_hash_counts)
-        and bool(bucket.lineage_hash_counts)
-        and bucket.pnl_basis == POST_COST_PNL_BASIS
-    )
-
-
-def _frontier_exact_replay_bucket(
-    *,
-    ledger_payload: Mapping[str, Any],
-    raw_rows: Sequence[object],
-    full_window_start: date | None,
-    full_window_end: date | None,
-) -> RuntimeLedgerBucket | None:
-    bucket_range = _frontier_exact_replay_bucket_range(
-        ledger_payload=ledger_payload,
-        full_window_start=full_window_start,
-        full_window_end=full_window_end,
-    )
-    if bucket_range is None:
-        return None
-    runtime_rows = _frontier_exact_replay_rows(ledger_payload, raw_rows)
-    if not runtime_rows:
-        return None
-    buckets = build_runtime_ledger_buckets(
-        runtime_rows,
-        bucket_ranges=[bucket_range],
-        require_order_lifecycle=True,
-    )
-    if len(buckets) != 1:
-        return None
-    bucket = buckets[0]
-    if not _frontier_exact_replay_bucket_has_authority(bucket):
-        return None
-    return bucket
-
-
-def _exact_replay_ledger_artifact_update(
-    *,
-    args: argparse.Namespace,
-    root: Path,
-    candidate_index: int,
-    candidate_id: str,
-    full_window_payload: Mapping[str, Any],
-    dataset_snapshot_id: str = "",
-    replay_lineage: Mapping[str, Any] | None = None,
-    candidate_evaluation_key: Mapping[str, Any] | None = None,
-    replay_tape_validation: Mapping[str, Any] | None = None,
-    candidate_search_key: str = "",
-    candidate_symbols: Sequence[str] = (),
-    full_window_start: date | None = None,
-    full_window_end: date | None = None,
-    proof_only_reason: str = "",
-) -> dict[str, Any]:
-    ledger_payload = _mapping(full_window_payload.get("exact_replay_ledger"))
-    raw_rows = ledger_payload.get("runtime_ledger_rows")
-    if not isinstance(raw_rows, list) or not raw_rows:
-        return {}
-    runtime_bucket = _frontier_exact_replay_bucket(
-        ledger_payload=ledger_payload,
-        raw_rows=cast(Sequence[object], raw_rows),
-        full_window_start=full_window_start,
-        full_window_end=full_window_end,
-    )
-    if runtime_bucket is None:
-        return {}
-    try:
-        fill_row_count = int(ledger_payload.get("fill_row_count") or 0)
-    except (TypeError, ValueError):
-        return {}
-    if fill_row_count != runtime_bucket.fill_count:
-        return {}
-    artifact_dir = _order_type_ablation_artifact_dir(args=args, root=root)
-    artifact_path = (
-        artifact_dir / f"candidate-{candidate_index:04d}-exact-replay-ledger.json"
-    )
-    artifact_ref = str(artifact_path)
-    artifact_payload = {
-        **ledger_payload,
-        "artifact_ref": artifact_ref,
-        "artifact_kind": "exact_replay_ledger",
-        "candidate_id": candidate_id,
-        "proof_authority": False,
-        "promotion_authority": False,
-        "promotion_proof": False,
-        "authority": "exact_replay_probation_only",
-        "authority_blockers": [
-            "source_backed_runtime_ledger_required",
-            "live_paper_runtime_evidence_required",
-        ],
-    }
-    if dataset_snapshot_id:
-        artifact_payload["dataset_snapshot_id"] = dataset_snapshot_id
-    if replay_lineage:
-        artifact_payload["replay_lineage"] = dict(replay_lineage)
-        artifact_payload["replay_lineage_hash"] = str(
-            replay_lineage.get("lineage_hash") or ""
-        )
-    if candidate_evaluation_key:
-        artifact_payload["candidate_evaluation_key_payload"] = dict(
-            candidate_evaluation_key
-        )
-        artifact_payload["candidate_evaluation_key"] = str(
-            candidate_evaluation_key.get("candidate_evaluation_key") or ""
-        )
-    if replay_tape_validation:
-        artifact_payload["replay_tape"] = {
-            **_replay_tape_selection_metadata(replay_tape_validation),
-            "status": str(replay_tape_validation.get("status") or ""),
-            "tape_path": str(replay_tape_validation.get("tape_path") or ""),
-            "manifest_path": str(replay_tape_validation.get("manifest_path") or ""),
-            "artifact_refs": dict(
-                cast(
-                    Mapping[str, Any], replay_tape_validation.get("artifact_refs") or {}
-                )
-            ),
-        }
-    if candidate_search_key:
-        artifact_payload["candidate_search_key"] = candidate_search_key
-    if candidate_symbols:
-        artifact_payload["candidate_symbols"] = list(candidate_symbols)
-    if full_window_start is not None and full_window_end is not None:
-        artifact_payload["full_window"] = {
-            "start_date": full_window_start.isoformat(),
-            "end_date": full_window_end.isoformat(),
-        }
-    if proof_only_reason:
-        artifact_payload["proof_only"] = True
-        artifact_payload["proof_only_reason"] = proof_only_reason
-    _write_json_output(artifact_path, artifact_payload)
-    update: dict[str, Any] = {
-        "exact_replay_ledger_artifact_ref": artifact_ref,
-        "exact_replay_ledger_artifact_row_count": len(raw_rows),
-        "exact_replay_ledger_artifact_fill_count": fill_row_count,
-        "runtime_ledger_closed_trade_count": runtime_bucket.closed_trade_count,
-        "runtime_ledger_open_position_count": runtime_bucket.open_position_count,
-        "runtime_ledger_filled_notional": str(runtime_bucket.filled_notional),
-        "runtime_ledger_net_strategy_pnl_after_costs": str(
-            runtime_bucket.net_strategy_pnl_after_costs
-        ),
-        "runtime_ledger_post_cost_expectancy_bps": str(
-            runtime_bucket.post_cost_expectancy_bps
-        ),
-        "runtime_ledger_cost_basis_counts": dict(runtime_bucket.cost_basis_counts),
-        "runtime_ledger_cost_basis_count": sum(
-            runtime_bucket.cost_basis_counts.values()
-        ),
-        "runtime_ledger_pnl_basis": POST_COST_PNL_BASIS,
-        "runtime_ledger_pnl_source": "exact_replay_runtime_ledger",
-        "exact_replay_ledger_artifact_proof_authority": False,
-        "promotion_authority": False,
-        "final_promotion_authority": False,
-        "promotion_authority_blockers": [
-            "source_backed_runtime_ledger_required",
-            "live_paper_runtime_evidence_required",
-        ],
-    }
-    if proof_only_reason:
-        update["exact_replay_ledger_artifact_proof_only"] = True
-        update["exact_replay_ledger_artifact_proof_only_reason"] = proof_only_reason
-    return update
-
-
-def _order_type_replay_arm_summary(
-    *,
-    order_type: str,
-    payload: Mapping[str, Any],
-) -> dict[str, Any]:
-    summary = summarize_replay_profitability(payload)
-    daily_filled_notional = _daily_filled_notional(payload)
-    total_filled_notional = sum(daily_filled_notional.values(), Decimal("0"))
-    avg_filled_notional_per_day = (
-        total_filled_notional / Decimal(summary.trading_day_count)
-        if summary.trading_day_count > 0
-        else Decimal("0")
-    )
-    arm_summary: dict[str, Any] = {
-        "order_type": order_type,
-        "start_date": summary.start_date,
-        "end_date": summary.end_date,
-        "trading_day_count": summary.trading_day_count,
-        "net_pnl": str(summary.net_pnl),
-        "net_per_day": str(summary.net_per_day),
-        "active_days": summary.active_days,
-        "decision_count": summary.decision_count,
-        "filled_count": summary.filled_count,
-        "sample_count": _forced_order_type_sample_count(payload, order_type=order_type),
-        "limit_fill_rate": str(payload.get("limit_fill_rate", "0")),
-        "total_filled_notional": str(total_filled_notional),
-        "avg_filled_notional_per_day": str(avg_filled_notional_per_day),
-        "payload_sha256": _payload_digest(payload),
-        "daily_net": {day: str(value) for day, value in summary.daily_net.items()},
-        "daily_filled_notional": {
-            day: str(value) for day, value in daily_filled_notional.items()
-        },
-    }
-    arm_summary.update(_order_type_execution_metrics(payload))
-    return arm_summary
-
-
-def _order_type_ablation_payload(
-    *,
-    candidate_index: int,
-    candidate_id: str,
-    policy: OrderTypeAblationPolicy,
-    candidate_params: Mapping[str, Any],
-    strategy_overrides: Mapping[str, Any],
-    market_payload: Mapping[str, Any],
-    limit_payload: Mapping[str, Any],
-    start_date: date,
-    end_date: date,
-) -> tuple[dict[str, Any], dict[str, Any]]:
-    selected_order_type = _selected_entry_order_type(
-        candidate_params=candidate_params,
-        strategy_overrides=strategy_overrides,
-    )
-    alternative_order_type = "limit" if selected_order_type == "market" else "market"
-    market_summary = _order_type_replay_arm_summary(
-        order_type="market",
-        payload=market_payload,
-    )
-    limit_summary = _order_type_replay_arm_summary(
-        order_type="limit",
-        payload=limit_payload,
-    )
-    selected_summary = (
-        market_summary if selected_order_type == "market" else limit_summary
-    )
-    alternative_summary = (
-        limit_summary if selected_order_type == "market" else market_summary
-    )
-    selected_net_per_day = Decimal(str(selected_summary["net_per_day"]))
-    alternative_net_per_day = Decimal(str(alternative_summary["net_per_day"]))
-    opportunity_cost_per_day = max(
-        Decimal("0"),
-        alternative_net_per_day - selected_net_per_day,
-    )
-    selected_notional = Decimal(str(selected_summary["avg_filled_notional_per_day"]))
-    alternative_notional = Decimal(
-        str(alternative_summary["avg_filled_notional_per_day"])
-    )
-    opportunity_cost_denominator = max(selected_notional, alternative_notional)
-    opportunity_cost_bps = (
-        opportunity_cost_per_day / opportunity_cost_denominator * Decimal("10000")
-        if opportunity_cost_denominator > 0
-        else Decimal("0")
-    )
-    market_sample_count = int(market_summary["sample_count"])
-    limit_sample_count = int(limit_summary["sample_count"])
-    sample_count = market_sample_count + limit_sample_count
-    passed = (
-        sample_count >= policy.min_sample_count
-        and opportunity_cost_bps <= policy.max_opportunity_cost_bps
-        and selected_net_per_day >= alternative_net_per_day
-    )
-    artifact_payload: dict[str, Any] = {
-        "schema_version": "torghut.order-type-ablation.v1",
-        "candidate_index": candidate_index,
-        "candidate_id": candidate_id,
-        "window": {
-            "start_date": start_date.isoformat(),
-            "end_date": end_date.isoformat(),
-        },
-        "policy": policy.to_payload(),
-        "candidate_params": dict(candidate_params),
-        "strategy_overrides": dict(strategy_overrides),
-        "selected_order_type": selected_order_type,
-        "alternative_order_type": alternative_order_type,
-        "market": market_summary,
-        "limit": limit_summary,
-        "market_sample_count": market_sample_count,
-        "limit_sample_count": limit_sample_count,
-        "sample_count": sample_count,
-        "selected_net_per_day": str(selected_net_per_day),
-        "alternative_net_per_day": str(alternative_net_per_day),
-        "opportunity_cost_per_day": str(opportunity_cost_per_day),
-        "opportunity_cost_denominator": str(opportunity_cost_denominator),
-        "opportunity_cost_bps": str(opportunity_cost_bps),
-        "passed": passed,
-    }
-    scorecard_update = {
-        "order_type_ablation_sample_count": sample_count,
-        "order_type_ablation_passed": passed,
-        "order_type_ablation_selected_order_type": selected_order_type,
-        "order_type_ablation_alternative_order_type": alternative_order_type,
-        "order_type_opportunity_cost_bps": str(opportunity_cost_bps),
-        "order_type_opportunity_cost_evidence_present": True,
-        "opportunity_cost_evidence_present": True,
-        "market_limit_order_mix_evidence_present": sample_count > 0,
-        "market_limit_order_mix_sample_count": sample_count,
-        "market_limit_execution_policy_passed": passed,
-        "limit_fill_probability_sample_count": limit_sample_count,
-        "limit_fill_probability_evidence_present": limit_sample_count > 0,
-    }
-    return artifact_payload, scorecard_update
-
-
-DELAY_ADJUSTED_DEPTH_STRESS_GRID_MS = (
-    Decimal("50"),
-    Decimal("150"),
-    Decimal("250"),
-)
-CONFORMAL_TAIL_RISK_ALPHA = Decimal("0.20")
-BREAKEVEN_TRANSACTION_COST_BUFFER_MIN_BPS = Decimal("1")
-MARKET_IMPACT_STRESS_SOURCE_MARKERS = (
-    "realistic_market_impact_arxiv_2603_29086_2026",
-    "double_square_root_impact_arxiv_2502_16246_2025",
-)
-
-
-def _p10(values: Sequence[Decimal]) -> Decimal:
-    if not values:
-        return Decimal("0")
-    ordered = sorted(values)
-    index = int((len(ordered) - 1) * 0.10)
-    return ordered[index]
-
-
-def _conformal_tail_loss_buffer(
-    daily_net: Mapping[str, Decimal],
-    *,
-    alpha: Decimal = CONFORMAL_TAIL_RISK_ALPHA,
-) -> Decimal:
-    losses = sorted((max(Decimal("0"), -value) for value in daily_net.values()))
-    if not losses:
-        return Decimal("0")
-    tail_fraction = max(Decimal("0"), min(Decimal("1"), alpha))
-    tail_count = max(
-        1,
-        int(
-            (Decimal(len(losses)) * tail_fraction).to_integral_value(
-                rounding=ROUND_CEILING
-            )
-        ),
-    )
-    tail_count = min(tail_count, len(losses))
-    return max(losses[-tail_count:], default=Decimal("0"))
-
-
-def _conformal_tail_risk_metrics(
-    *,
-    target_net_per_day: Decimal,
-    net_per_day: Decimal,
-    daily_net: Mapping[str, Decimal],
-) -> dict[str, Any]:
-    buffer_per_day = _conformal_tail_loss_buffer(daily_net)
-    adjusted_net_per_day = net_per_day - buffer_per_day
-    sample_count = len(daily_net)
-    passed = sample_count > 0 and adjusted_net_per_day >= target_net_per_day
-    return {
-        "conformal_tail_risk_required": True,
-        "conformal_tail_risk_model": "empirical_daily_loss_conformal_buffer",
-        "conformal_tail_risk_alpha": str(CONFORMAL_TAIL_RISK_ALPHA),
-        "conformal_tail_risk_sample_count": sample_count,
-        "conformal_tail_risk_buffer_per_day": str(buffer_per_day),
-        "conformal_tail_risk_adjusted_net_pnl_per_day": str(adjusted_net_per_day),
-        "conformal_tail_risk_target_net_pnl_per_day": str(target_net_per_day),
-        "conformal_tail_risk_passed": passed,
-        "conformal_tail_risk_source_markers": [
-            "regime_weighted_conformal_var_arxiv_2602_03903_2026"
-        ],
-    }
-
-
-def _breakeven_transaction_cost_buffer_metrics(
-    *,
-    target_net_per_day: Decimal,
-    conformal_adjusted_net_per_day: Decimal,
-    avg_filled_notional_per_day: Decimal,
-    required_cost_buffer_bps: Decimal,
-) -> dict[str, Any]:
-    selected_cost_buffer_bps = max(
-        BREAKEVEN_TRANSACTION_COST_BUFFER_MIN_BPS,
-        required_cost_buffer_bps,
-    )
-    breakeven_surplus_per_day = max(
-        Decimal("0"),
-        conformal_adjusted_net_per_day - target_net_per_day,
-    )
-    breakeven_buffer_bps = (
-        breakeven_surplus_per_day / avg_filled_notional_per_day * Decimal("10000")
-        if avg_filled_notional_per_day > 0
-        else Decimal("0")
-    )
-    buffer_cost_per_day = (
-        avg_filled_notional_per_day * selected_cost_buffer_bps / Decimal("10000")
-    )
-    buffered_net_per_day = conformal_adjusted_net_per_day - buffer_cost_per_day
-    passed = (
-        avg_filled_notional_per_day > 0
-        and breakeven_buffer_bps >= selected_cost_buffer_bps
-        and buffered_net_per_day >= target_net_per_day
-    )
-    return {
-        "required_breakeven_transaction_cost_buffer": True,
-        "required_seed_model_family_robustness": True,
-        "breakeven_transaction_cost_buffer_passed": passed,
-        "breakeven_transaction_cost_buffer_bps": str(breakeven_buffer_bps),
-        "transaction_cost_buffer_bps": str(selected_cost_buffer_bps),
-        "transaction_cost_buffer_cost_per_day": str(buffer_cost_per_day),
-        "post_cost_net_pnl_after_breakeven_transaction_cost_buffer": str(
-            buffered_net_per_day
-        ),
-        "breakeven_transaction_cost_buffer_target_net_pnl_per_day": str(
-            target_net_per_day
-        ),
-        "seed_model_family_robustness_status": (
-            "required_not_materialized_by_single_frontier_replay"
-        ),
-        "seed_robustness_passed": False,
-        "seed_robustness_sample_count": 0,
-        "model_family_robustness_passed": False,
-        "model_family_robustness_family_count": 0,
-        "breakeven_transaction_cost_buffer_source_markers": [
-            "regime_weighted_conformal_var_arxiv_2602_03903_2026",
-            "realistic_market_impact_arxiv_2603_29086_2026",
-        ],
-        "seed_model_family_robustness_source_markers": [
-            "regime_weighted_conformal_var_arxiv_2602_03903_2026"
-        ],
-    }
-
-
-def _delay_depth_fillability(
-    *,
-    daily_filled_notional: Mapping[str, Decimal],
-    daily_liquidity_notional: Mapping[str, Decimal],
-    stress_ms: Decimal,
-) -> tuple[Decimal, int, list[Decimal]]:
-    haircut_rate = min(Decimal("0.50"), stress_ms / Decimal("1000"))
-    total_fillable_notional = Decimal("0")
-    missing_liquidity_day_count = 0
-    active_day_fillable: list[Decimal] = []
-    for day, filled_notional in daily_filled_notional.items():
-        if filled_notional <= 0:
-            continue
-        liquidity_notional = daily_liquidity_notional.get(day, Decimal("0"))
-        if liquidity_notional <= 0:
-            missing_liquidity_day_count += 1
-            active_day_fillable.append(Decimal("0"))
-            continue
-        fillable_notional = min(
-            filled_notional,
-            liquidity_notional * (Decimal("1") - haircut_rate),
-        )
-        active_day_fillable.append(fillable_notional)
-        total_fillable_notional += fillable_notional
-    return total_fillable_notional, missing_liquidity_day_count, active_day_fillable
-
-
-def _implementation_uncertainty_metrics(
-    *,
-    target_net_per_day: Decimal,
-    net_per_day: Decimal,
-    avg_filled_notional_per_day: Decimal,
-    square_root_impact_cost_bps: Decimal,
-    almgren_chriss_impact_cost_bps: Decimal,
-    nonlinear_impact_cost_bps: Decimal,
-    delay_depth_net_per_day: Decimal,
-) -> dict[str, Any]:
-    impact_scenarios = {
-        "square_root": net_per_day
-        - (
-            avg_filled_notional_per_day * square_root_impact_cost_bps / Decimal("10000")
-        ),
-        "almgren_chriss_proxy": net_per_day
-        - (
-            avg_filled_notional_per_day
-            * almgren_chriss_impact_cost_bps
-            / Decimal("10000")
-        ),
-        "selected_nonlinear_impact": net_per_day
-        - (avg_filled_notional_per_day * nonlinear_impact_cost_bps / Decimal("10000")),
-        "impact_decay_reversion_1_5x": net_per_day
-        - (
-            avg_filled_notional_per_day
-            * nonlinear_impact_cost_bps
-            * Decimal("1.5")
-            / Decimal("10000")
-        ),
-        "latency_depth_fillability": delay_depth_net_per_day,
-    }
-    lower_bound = min(impact_scenarios.values(), default=Decimal("0"))
-    upper_bound = max(impact_scenarios.values(), default=Decimal("0"))
-    interval_width = upper_bound - lower_bound
-    passed = (
-        bool(impact_scenarios)
-        and avg_filled_notional_per_day > 0
-        and lower_bound >= target_net_per_day
-    )
-    return {
-        "implementation_uncertainty_required": True,
-        "implementation_uncertainty_model": "impact_latency_cost_model_interval",
-        "implementation_uncertainty_model_count": len(impact_scenarios),
-        "implementation_uncertainty_stability_passed": passed,
-        "implementation_uncertainty_lower_net_pnl_per_day": str(lower_bound),
-        "implementation_uncertainty_upper_net_pnl_per_day": str(upper_bound),
-        "implementation_uncertainty_interval_width_per_day": str(interval_width),
-        "implementation_uncertainty_target_net_pnl_per_day": str(target_net_per_day),
-        "implementation_uncertainty_scenarios": {
-            name: str(value) for name, value in impact_scenarios.items()
-        },
-        "implementation_uncertainty_source_markers": [
-            "lob_simulation_reality_gap_arxiv_2603_24137_2026",
-            "order_flow_market_impact_volatility_arxiv_2601_23172_2026",
-            "double_square_root_impact_arxiv_2502_16246_2025",
-            "implementation_risk_backtesting_arxiv_2603_20319_2026",
-        ],
-    }
-
-
-def _replay_stress_metrics(
-    *,
-    target_net_per_day: Decimal,
-    net_per_day: Decimal,
-    trading_day_count: int,
-    daily_filled_notional: Mapping[str, Decimal],
-    daily_liquidity_notional: Mapping[str, Decimal],
-    avg_filled_notional_per_day: Decimal,
-    total_filled_notional: Decimal,
-    total_liquidity_notional: Decimal,
-    fill_survival_rate: Decimal | None = None,
-    fill_survival_sample_count: int = 0,
-    queue_ratio_p95: Decimal | None = None,
-    queue_ahead_depletion_evidence_present: bool = False,
-    queue_ahead_depletion_sample_count: int = 0,
-) -> dict[str, Any]:
-    participation = (
-        total_filled_notional / total_liquidity_notional
-        if total_filled_notional > 0 and total_liquidity_notional > 0
-        else Decimal("0")
-    )
-    square_root_impact_cost_bps = (
-        participation.sqrt() * Decimal("100") if participation > 0 else Decimal("0")
-    )
-    almgren_chriss_temporary_impact_bps = participation * Decimal("125")
-    almgren_chriss_permanent_impact_bps = (
-        participation.sqrt() * Decimal("25") if participation > 0 else Decimal("0")
-    )
-    almgren_chriss_impact_cost_bps = (
-        almgren_chriss_temporary_impact_bps + almgren_chriss_permanent_impact_bps
-    )
-    nonlinear_impact_cost_bps = max(
-        Decimal("1"),
-        square_root_impact_cost_bps,
-        almgren_chriss_impact_cost_bps,
-    )
-    market_impact_model = (
-        "almgren_chriss_proxy"
-        if almgren_chriss_impact_cost_bps > square_root_impact_cost_bps
-        else "square_root"
-    )
-    market_impact_net_per_day = net_per_day - (
-        avg_filled_notional_per_day * nonlinear_impact_cost_bps / Decimal("10000")
-    )
-    (
-        delay_depth_total_fillable_notional,
-        delay_depth_missing_liquidity_day_count,
-        _active_day_fillable,
-    ) = _delay_depth_fillability(
-        daily_filled_notional=daily_filled_notional,
-        daily_liquidity_notional=daily_liquidity_notional,
-        stress_ms=Decimal("50"),
-    )
-    grid_fillability = {
-        str(stress_ms): _delay_depth_fillability(
-            daily_filled_notional=daily_filled_notional,
-            daily_liquidity_notional=daily_liquidity_notional,
-            stress_ms=stress_ms,
-        )
-        for stress_ms in DELAY_ADJUSTED_DEPTH_STRESS_GRID_MS
-    }
-    max_grid_stress_ms = max(DELAY_ADJUSTED_DEPTH_STRESS_GRID_MS)
-    (
-        grid_worst_total_fillable_notional,
-        grid_worst_missing_liquidity_day_count,
-        grid_worst_active_day_fillable,
-    ) = grid_fillability[str(max_grid_stress_ms)]
-    p10_active_day_fillable = _p10(grid_worst_active_day_fillable)
-    worst_active_day_fillable = (
-        min(grid_worst_active_day_fillable)
-        if grid_worst_active_day_fillable
-        else Decimal("0")
-    )
-    tail_coverage_passed = (
-        bool(grid_worst_active_day_fillable)
-        and grid_worst_missing_liquidity_day_count == 0
-        and p10_active_day_fillable > 0
-        and worst_active_day_fillable > 0
-    )
-    delay_depth_fillable_notional_per_day = (
-        delay_depth_total_fillable_notional / Decimal(trading_day_count)
-        if trading_day_count > 0
-        else Decimal("0")
-    )
-    delay_depth_fillable_ratio = (
-        delay_depth_total_fillable_notional / total_filled_notional
-        if total_filled_notional > 0
-        else Decimal("1")
-    )
-    survival_adjusted_fillable_ratio = delay_depth_fillable_ratio
-    if fill_survival_sample_count > 0 and fill_survival_rate is not None:
-        survival_adjusted_fillable_ratio *= max(
-            Decimal("0"), min(Decimal("1"), fill_survival_rate)
-        )
-    delay_depth_net_per_day = (net_per_day * survival_adjusted_fillable_ratio) - (
-        delay_depth_fillable_notional_per_day * Decimal("1") / Decimal("10000")
-    )
-    queue_position_nonfill_opportunity_cost_per_day = max(
-        Decimal("0"), net_per_day - delay_depth_net_per_day
-    )
-    queue_position_nonfill_opportunity_cost_bps = (
-        queue_position_nonfill_opportunity_cost_per_day
-        / avg_filled_notional_per_day
-        * Decimal("10000")
-        if avg_filled_notional_per_day > 0
-        else Decimal("0")
-    )
-    implementation_uncertainty = _implementation_uncertainty_metrics(
-        target_net_per_day=target_net_per_day,
-        net_per_day=net_per_day,
-        avg_filled_notional_per_day=avg_filled_notional_per_day,
-        square_root_impact_cost_bps=square_root_impact_cost_bps,
-        almgren_chriss_impact_cost_bps=almgren_chriss_impact_cost_bps,
-        nonlinear_impact_cost_bps=nonlinear_impact_cost_bps,
-        delay_depth_net_per_day=delay_depth_net_per_day,
-    )
-    return {
-        "market_impact_stress_passed": bool(
-            total_liquidity_notional > 0
-            and avg_filled_notional_per_day > 0
-            and market_impact_net_per_day > 0
-        ),
-        "market_impact_stress_model": market_impact_model,
-        "market_impact_stress_cost_bps": str(nonlinear_impact_cost_bps),
-        "market_impact_stress_net_pnl_per_day": str(market_impact_net_per_day),
-        "market_impact_stress_source_markers": list(
-            MARKET_IMPACT_STRESS_SOURCE_MARKERS
-        ),
-        "market_impact_stress_components": {
-            "square_root_cost_bps": str(square_root_impact_cost_bps),
-            "almgren_chriss_temporary_impact_bps": str(
-                almgren_chriss_temporary_impact_bps
-            ),
-            "almgren_chriss_permanent_impact_bps": str(
-                almgren_chriss_permanent_impact_bps
-            ),
-            "almgren_chriss_cost_bps": str(almgren_chriss_impact_cost_bps),
-            "selected_cost_bps": str(nonlinear_impact_cost_bps),
-            "selected_model": market_impact_model,
-            "source_marker": "realistic_market_impact_arxiv_2603_29086_2026",
-            "source_markers": list(MARKET_IMPACT_STRESS_SOURCE_MARKERS),
-        },
-        "nonlinear_market_impact_stress_passed": bool(
-            total_liquidity_notional > 0
-            and avg_filled_notional_per_day > 0
-            and market_impact_net_per_day > 0
-        ),
-        "nonlinear_market_impact_stress_model": market_impact_model,
-        "nonlinear_market_impact_stress_cost_bps": str(nonlinear_impact_cost_bps),
-        "nonlinear_market_impact_stress_net_pnl_per_day": str(
-            market_impact_net_per_day
-        ),
-        "permanent_impact_decay_model": "exponential_decay_proxy",
-        "delay_adjusted_depth_stress_passed": bool(
-            total_liquidity_notional > 0
-            and grid_worst_missing_liquidity_day_count == 0
-            and tail_coverage_passed
-            and delay_depth_fillable_notional_per_day > 0
-            and delay_depth_net_per_day > 0
-        ),
-        "delay_adjusted_depth_stress_model": "latency_depth_haircut",
-        "delay_adjusted_depth_stress_ms": "50",
-        "delay_adjusted_depth_latency_grid_ms": [
-            str(stress_ms) for stress_ms in DELAY_ADJUSTED_DEPTH_STRESS_GRID_MS
-        ],
-        "delay_adjusted_depth_grid_max_stress_ms": str(max_grid_stress_ms),
-        "delay_adjusted_depth_liquidity_evidence_present": bool(
-            total_liquidity_notional > 0 and grid_worst_missing_liquidity_day_count == 0
-        ),
-        "delay_adjusted_depth_liquidity_missing_day_count": max(
-            delay_depth_missing_liquidity_day_count,
-            grid_worst_missing_liquidity_day_count,
-        ),
-        "delay_adjusted_depth_fillable_notional_per_day": str(
-            delay_depth_fillable_notional_per_day
-        ),
-        "delay_adjusted_depth_worst_grid_fillable_notional_per_day": str(
-            grid_worst_total_fillable_notional / Decimal(trading_day_count)
-            if trading_day_count > 0
-            else Decimal("0")
-        ),
-        "delay_adjusted_depth_worst_active_day_fillable_notional": str(
-            worst_active_day_fillable
-        ),
-        "delay_adjusted_depth_p10_active_day_fillable_notional": str(
-            p10_active_day_fillable
-        ),
-        "delay_adjusted_depth_tail_coverage_passed": tail_coverage_passed,
-        "delay_adjusted_depth_fillable_ratio": str(delay_depth_fillable_ratio),
-        "delay_adjusted_depth_survival_adjusted_fillable_ratio": str(
-            survival_adjusted_fillable_ratio
-        ),
-        "delay_adjusted_depth_fill_survival_evidence_present": (
-            fill_survival_sample_count > 0
-        ),
-        "delay_adjusted_depth_fill_survival_sample_count": fill_survival_sample_count,
-        "delay_adjusted_depth_fill_survival_rate": str(fill_survival_rate)
-        if fill_survival_rate is not None
-        else "",
-        "delay_adjusted_depth_queue_ratio_p95": str(queue_ratio_p95)
-        if queue_ratio_p95 is not None
-        else "",
-        "queue_position_survival_fill_curve_evidence_present": (
-            fill_survival_sample_count > 0
-            and queue_ratio_p95 is not None
-            and queue_ahead_depletion_evidence_present
-            and queue_ahead_depletion_sample_count > 0
-        ),
-        "queue_position_survival_sample_count": fill_survival_sample_count,
-        "queue_position_survival_fill_rate": str(fill_survival_rate)
-        if fill_survival_rate is not None
-        else "",
-        "queue_position_survival_queue_ratio_p95": str(queue_ratio_p95)
-        if queue_ratio_p95 is not None
-        else "",
-        "queue_position_survival_queue_ahead_depletion_evidence_present": (
-            queue_ahead_depletion_evidence_present
-        ),
-        "queue_position_survival_queue_ahead_depletion_sample_count": (
-            queue_ahead_depletion_sample_count
-        ),
-        "queue_position_survival_adjusted_fillable_ratio": str(
-            survival_adjusted_fillable_ratio
-        ),
-        "queue_position_survival_nonfill_opportunity_cost_per_day": str(
-            queue_position_nonfill_opportunity_cost_per_day
-        ),
-        "queue_position_survival_nonfill_opportunity_cost_bps": str(
-            queue_position_nonfill_opportunity_cost_bps
-        ),
-        "queue_position_survival_stress_net_pnl_per_day": str(delay_depth_net_per_day),
-        "post_cost_net_pnl_after_queue_position_survival_fill_stress": str(
-            delay_depth_net_per_day
-        ),
-        "queue_position_survival_source_marker": (
-            "queue_position_survival_fill_probability_arxiv_2512_05734_2025"
-        ),
-        "delay_adjusted_depth_unfillable_notional_per_day": str(
-            max(
-                Decimal("0"),
-                (total_filled_notional - delay_depth_total_fillable_notional)
-                / Decimal(trading_day_count)
-                if trading_day_count > 0
-                else Decimal("0"),
-            )
-        ),
-        "delay_adjusted_depth_stress_net_pnl_per_day": str(delay_depth_net_per_day),
-        **implementation_uncertainty,
-    }
-
-
-def _decimal_payload_metric(
-    payload: Mapping[str, Any],
-    key: str,
-    *,
-    default: Decimal,
-) -> Decimal:
-    raw_value = payload.get(key)
-    if raw_value is None:
-        return default
-    return Decimal(str(raw_value))
-
-
-def _max_best_day_share_of_total_pnl(
-    *,
-    daily_net: Mapping[str, Decimal],
-    total_net_pnl: Decimal,
-) -> Decimal:
-    if total_net_pnl <= 0:
-        return Decimal("1")
-    best_positive_day = max(
-        (value for value in daily_net.values() if value > 0), default=Decimal("0")
-    )
-    if best_positive_day <= 0:
-        return Decimal("0")
-    return best_positive_day / total_net_pnl
-
-
-def _consistency_penalty(
-    *,
-    full_window_payload: Mapping[str, Any],
-    policy: FullWindowConsistencyPolicy,
-) -> tuple[Decimal, dict[str, Any]]:
-    summary = summarize_replay_profitability(full_window_payload)
-    daily_filled_notional = _daily_filled_notional(full_window_payload)
-    daily_liquidity_notional = _daily_liquidity_notional(full_window_payload)
-    daily_gross_exposure_pct_equity = _daily_decimal_metric(
-        full_window_payload,
-        "max_gross_exposure_pct_equity",
-    )
-    daily_min_cash = _daily_decimal_metric(full_window_payload, "min_cash")
-    daily_negative_cash_observations = _daily_int_metric(
-        full_window_payload,
-        "negative_cash_observation_count",
-    )
-    negative_days = sum(1 for value in summary.daily_net.values() if value < 0)
-    positive_days = sum(1 for value in summary.daily_net.values() if value > 0)
-    drawdown = _max_drawdown_from_daily_net(summary.daily_net)
-    active_ratio = (
-        Decimal(summary.active_days) / Decimal(summary.trading_day_count)
-        if summary.trading_day_count > 0
-        else Decimal("0")
-    )
-    total_filled_notional = sum(daily_filled_notional.values(), Decimal("0"))
-    total_liquidity_notional = sum(
-        daily_liquidity_notional.values(),
-        Decimal("0"),
-    )
-    avg_filled_notional_per_day = (
-        total_filled_notional / Decimal(summary.trading_day_count)
-        if summary.trading_day_count > 0
-        else Decimal("0")
-    )
-    avg_liquidity_notional_per_day = (
-        total_liquidity_notional / Decimal(summary.trading_day_count)
-        if summary.trading_day_count > 0
-        else Decimal("0")
-    )
-    avg_filled_notional_per_active_day = (
-        total_filled_notional / Decimal(summary.active_days)
-        if summary.active_days > 0
-        else Decimal("0")
-    )
-    order_lifecycle_metrics = _order_lifecycle_metrics(full_window_payload)
-    fill_survival_rate = _optional_decimal(
-        order_lifecycle_metrics.get("fill_survival_fill_rate")
-    )
-    queue_ratio_p95 = _optional_decimal(
-        order_lifecycle_metrics.get("order_qty_to_touch_qty_ratio_p95")
-    )
-    replay_stress_metrics = _replay_stress_metrics(
-        target_net_per_day=policy.target_net_per_day,
-        net_per_day=summary.net_per_day,
-        trading_day_count=summary.trading_day_count,
-        daily_filled_notional=daily_filled_notional,
-        daily_liquidity_notional=daily_liquidity_notional,
-        avg_filled_notional_per_day=avg_filled_notional_per_day,
-        total_filled_notional=total_filled_notional,
-        total_liquidity_notional=total_liquidity_notional,
-        fill_survival_rate=fill_survival_rate,
-        fill_survival_sample_count=int(
-            order_lifecycle_metrics.get("fill_survival_sample_count") or 0
-        ),
-        queue_ratio_p95=queue_ratio_p95,
-        queue_ahead_depletion_evidence_present=bool(
-            order_lifecycle_metrics.get("queue_ahead_depletion_evidence_present")
-        ),
-        queue_ahead_depletion_sample_count=int(
-            order_lifecycle_metrics.get("queue_ahead_depletion_sample_count") or 0
-        ),
-    )
-    conformal_tail_risk_metrics = _conformal_tail_risk_metrics(
-        target_net_per_day=policy.target_net_per_day,
-        net_per_day=summary.net_per_day,
-        daily_net=summary.daily_net,
-    )
-    breakeven_cost_buffer_metrics = _breakeven_transaction_cost_buffer_metrics(
-        target_net_per_day=policy.target_net_per_day,
-        conformal_adjusted_net_per_day=Decimal(
-            str(
-                conformal_tail_risk_metrics[
-                    "conformal_tail_risk_adjusted_net_pnl_per_day"
-                ]
-            )
-        ),
-        avg_filled_notional_per_day=avg_filled_notional_per_day,
-        required_cost_buffer_bps=Decimal(
-            str(replay_stress_metrics.get("market_impact_stress_cost_bps") or "0")
-        ),
-    )
-    best_day_share_of_total_pnl = _max_best_day_share_of_total_pnl(
-        daily_net=summary.daily_net,
-        total_net_pnl=summary.net_pnl,
-    )
-    min_daily_net_pnl = min(summary.daily_net.values(), default=Decimal("0"))
-    daily_net_below_min_count = sum(
-        1 for value in summary.daily_net.values() if value < policy.min_daily_net_pnl
-    )
-    if (
-        policy.min_daily_net_pnl > 0
-        and len(summary.daily_net) < summary.trading_day_count
-    ):
-        daily_net_below_min_count += summary.trading_day_count - len(summary.daily_net)
-    max_gross_exposure_pct_equity = max(
-        (
-            _decimal_payload_metric(
-                full_window_payload,
-                "max_gross_exposure_pct_equity",
-                default=Decimal("0"),
-            ),
-            *daily_gross_exposure_pct_equity.values(),
-        ),
-        default=Decimal("0"),
-    )
-    min_cash = min(
-        (
-            _decimal_payload_metric(
-                full_window_payload, "min_cash", default=Decimal("0")
-            ),
-            *daily_min_cash.values(),
-        ),
-        default=Decimal("0"),
-    )
-    negative_cash_observation_count = max(
-        int(full_window_payload.get("negative_cash_observation_count") or 0),
-        sum(daily_negative_cash_observations.values()),
-    )
-    penalties = Decimal("0")
-
-    if (
-        policy.min_window_weekday_count > 0
-        and summary.trading_day_count < policy.min_window_weekday_count
-    ):
-        penalties += Decimal(
-            policy.min_window_weekday_count - summary.trading_day_count
-        ) * Decimal("1000")
-    if summary.net_per_day < policy.target_net_per_day:
-        penalties += policy.target_net_per_day - summary.net_per_day
-    if policy.min_daily_net_pnl > 0 and daily_net_below_min_count > 0:
-        penalties += sum(
-            max(Decimal("0"), policy.min_daily_net_pnl - value)
-            for value in summary.daily_net.values()
-        )
-        penalties += (
-            Decimal(max(0, summary.trading_day_count - len(summary.daily_net)))
-            * policy.min_daily_net_pnl
-        )
-    effective_min_active_days = min(policy.min_active_days, summary.trading_day_count)
-    if summary.active_days < effective_min_active_days:
-        penalties += Decimal(effective_min_active_days - summary.active_days) * Decimal(
-            "250"
-        )
-    if active_ratio < policy.min_active_ratio:
-        penalties += (policy.min_active_ratio - active_ratio) * Decimal("2000")
-    effective_min_positive_days = min(
-        policy.min_positive_days,
-        summary.trading_day_count,
-    )
-    if positive_days < effective_min_positive_days:
-        penalties += Decimal(effective_min_positive_days - positive_days) * Decimal(
-            "350"
-        )
-    if (
-        policy.require_every_day_active
-        and summary.active_days < summary.trading_day_count
-    ):
-        penalties += Decimal(summary.trading_day_count - summary.active_days) * Decimal(
-            "400"
-        )
-    if summary.worst_day_net < -policy.max_worst_day_loss:
-        penalties += abs(summary.worst_day_net + policy.max_worst_day_loss)
-    if negative_days > policy.max_negative_days:
-        penalties += Decimal(negative_days - policy.max_negative_days) * Decimal("300")
-    if drawdown > policy.max_drawdown:
-        penalties += drawdown - policy.max_drawdown
-    if best_day_share_of_total_pnl > policy.max_best_day_share_of_total_pnl:
-        penalties += (
-            best_day_share_of_total_pnl - policy.max_best_day_share_of_total_pnl
-        ) * abs(summary.net_pnl)
-    if avg_filled_notional_per_day < policy.min_avg_filled_notional_per_day:
-        penalties += (
-            policy.min_avg_filled_notional_per_day - avg_filled_notional_per_day
-        ) / Decimal("1000")
-    if (
-        avg_filled_notional_per_active_day
-        < policy.min_avg_filled_notional_per_active_day
-    ):
-        penalties += (
-            policy.min_avg_filled_notional_per_active_day
-            - avg_filled_notional_per_active_day
-        ) / Decimal("1000")
-    if max_gross_exposure_pct_equity > policy.max_gross_exposure_pct_equity:
-        penalties += (
-            max_gross_exposure_pct_equity - policy.max_gross_exposure_pct_equity
-        ) * Decimal("1000")
-    if min_cash < policy.min_cash:
-        penalties += policy.min_cash - min_cash
-    implementation_uncertainty_lower_bound = Decimal(
-        str(
-            replay_stress_metrics.get(
-                "implementation_uncertainty_lower_net_pnl_per_day"
-            )
-            or "0"
-        )
-    )
-    if not bool(
-        replay_stress_metrics.get("implementation_uncertainty_stability_passed")
-    ):
-        penalties += max(
-            Decimal("0"),
-            policy.target_net_per_day - implementation_uncertainty_lower_bound,
-        )
-    if not bool(conformal_tail_risk_metrics["conformal_tail_risk_passed"]):
-        penalties += max(
-            Decimal("0"),
-            policy.target_net_per_day
-            - Decimal(
-                str(
-                    conformal_tail_risk_metrics[
-                        "conformal_tail_risk_adjusted_net_pnl_per_day"
-                    ]
-                )
-            ),
-        )
-    if not bool(
-        breakeven_cost_buffer_metrics["breakeven_transaction_cost_buffer_passed"]
-    ):
-        penalties += max(
-            Decimal("0"),
-            policy.target_net_per_day
-            - Decimal(
-                str(
-                    breakeven_cost_buffer_metrics[
-                        "post_cost_net_pnl_after_breakeven_transaction_cost_buffer"
-                    ]
-                )
-            ),
-        )
-
-    return (
-        penalties,
-        {
-            "start_date": summary.start_date,
-            "end_date": summary.end_date,
-            "trading_day_count": summary.trading_day_count,
-            "min_window_weekday_count_required": policy.min_window_weekday_count,
-            "net_pnl": str(summary.net_pnl),
-            "net_per_day": str(summary.net_per_day),
-            "min_daily_net_pnl": str(min_daily_net_pnl),
-            "min_daily_net_pnl_required": str(policy.min_daily_net_pnl),
-            "daily_net_below_min_count": daily_net_below_min_count,
-            "active_days": summary.active_days,
-            "active_ratio": str(active_ratio),
-            "positive_days": positive_days,
-            "worst_day_net": str(summary.worst_day_net),
-            "negative_days": negative_days,
-            "max_drawdown": str(drawdown),
-            "total_filled_notional": str(total_filled_notional),
-            "avg_filled_notional_per_day": str(avg_filled_notional_per_day),
-            "avg_filled_notional_per_active_day": str(
-                avg_filled_notional_per_active_day
-            ),
-            "market_impact_liquidity_evidence_present": bool(daily_liquidity_notional),
-            "market_impact_liquidity_day_count": len(daily_liquidity_notional),
-            "market_impact_liquidity_missing_day_count": max(
-                0,
-                summary.trading_day_count - len(daily_liquidity_notional),
-            ),
-            "total_liquidity_notional": str(total_liquidity_notional),
-            "avg_liquidity_notional_per_day": str(avg_liquidity_notional_per_day),
-            **replay_stress_metrics,
-            **conformal_tail_risk_metrics,
-            **breakeven_cost_buffer_metrics,
-            "best_day_share_of_total_pnl": str(best_day_share_of_total_pnl),
-            "max_gross_exposure_pct_equity": str(max_gross_exposure_pct_equity),
-            "max_gross_exposure_pct_equity_required": str(
-                policy.max_gross_exposure_pct_equity
-            ),
-            "min_cash": str(min_cash),
-            "min_cash_required": str(policy.min_cash),
-            "negative_cash_observation_count": negative_cash_observation_count,
-            "daily_net": {day: str(value) for day, value in summary.daily_net.items()},
-            "daily_filled_notional": {
-                day: str(value) for day, value in daily_filled_notional.items()
-            },
-            "daily_liquidity_notional": {
-                day: str(value) for day, value in daily_liquidity_notional.items()
-            },
-            "daily_max_gross_exposure_pct_equity": {
-                day: str(value)
-                for day, value in daily_gross_exposure_pct_equity.items()
-            },
-            "daily_min_cash": {
-                day: str(value) for day, value in daily_min_cash.items()
-            },
-            "daily_negative_cash_observation_count": daily_negative_cash_observations,
-            **_order_type_execution_metrics(full_window_payload),
-            **order_lifecycle_metrics,
-        },
-    )
-
-
-def _second_oos_summary(
-    *,
-    second_oos_payload: Mapping[str, Any],
-    policy: FullWindowConsistencyPolicy,
-) -> tuple[Decimal, dict[str, Any]]:
-    summary = summarize_replay_profitability(second_oos_payload)
-    daily_filled_notional = _daily_filled_notional(second_oos_payload)
-    daily_liquidity_notional = _daily_liquidity_notional(second_oos_payload)
-    active_ratio = (
-        Decimal(summary.active_days) / Decimal(summary.trading_day_count)
-        if summary.trading_day_count > 0
-        else Decimal("0")
-    )
-    positive_days = sum(1 for value in summary.daily_net.values() if value > 0)
-    positive_ratio = (
-        Decimal(positive_days) / Decimal(summary.trading_day_count)
-        if summary.trading_day_count > 0
-        else Decimal("0")
-    )
-    drawdown = _max_drawdown_from_daily_net(summary.daily_net)
-    total_filled_notional = sum(daily_filled_notional.values(), Decimal("0"))
-    total_liquidity_notional = sum(
-        daily_liquidity_notional.values(),
-        Decimal("0"),
-    )
-    avg_filled_notional_per_day = (
-        total_filled_notional / Decimal(summary.trading_day_count)
-        if summary.trading_day_count > 0
-        else Decimal("0")
-    )
-    avg_liquidity_notional_per_day = (
-        total_liquidity_notional / Decimal(summary.trading_day_count)
-        if summary.trading_day_count > 0
-        else Decimal("0")
-    )
-    reasons: list[str] = []
-    penalty = Decimal("0")
-    if summary.trading_day_count <= 0:
-        reasons.append(f"{_SECOND_OOS_WINDOW_ID}_trading_days_missing")
-    if summary.decision_count <= 0:
-        reasons.append(f"{_SECOND_OOS_WINDOW_ID}_no_decisions")
-    if summary.filled_count <= 0:
-        reasons.append(f"{_SECOND_OOS_WINDOW_ID}_no_fills")
-    if summary.net_per_day < policy.target_net_per_day:
-        reasons.append(f"{_SECOND_OOS_WINDOW_ID}_net_per_day_below_target")
-        penalty += policy.target_net_per_day - summary.net_per_day
-    if active_ratio < policy.min_active_ratio:
-        reasons.append(f"{_SECOND_OOS_WINDOW_ID}_active_ratio_below_min")
-        penalty += (policy.min_active_ratio - active_ratio) * Decimal("2000")
-    if drawdown > policy.max_drawdown:
-        reasons.append(f"{_SECOND_OOS_WINDOW_ID}_max_drawdown_above_max")
-        penalty += drawdown - policy.max_drawdown
-    if summary.worst_day_net < -policy.max_worst_day_loss:
-        reasons.append(f"{_SECOND_OOS_WINDOW_ID}_worst_day_loss_above_max")
-        penalty += abs(summary.worst_day_net + policy.max_worst_day_loss)
-    if avg_filled_notional_per_day < policy.min_avg_filled_notional_per_day:
-        reasons.append(f"{_SECOND_OOS_WINDOW_ID}_filled_notional_below_min")
-        penalty += (
-            policy.min_avg_filled_notional_per_day - avg_filled_notional_per_day
-        ) / Decimal("1000")
-
-    return (
-        penalty,
-        {
-            "schema_version": "torghut.frontier-second-oos-window.v1",
-            "window_id": _SECOND_OOS_WINDOW_ID,
-            "start_date": summary.start_date,
-            "end_date": summary.end_date,
-            "trading_day_count": summary.trading_day_count,
-            "net_pnl": str(summary.net_pnl),
-            "net_per_day": str(summary.net_per_day),
-            "target_net_per_day": str(policy.target_net_per_day),
-            "active_days": summary.active_days,
-            "active_ratio": str(active_ratio),
-            "positive_days": positive_days,
-            "positive_ratio": str(positive_ratio),
-            "decision_count": summary.decision_count,
-            "filled_count": summary.filled_count,
-            "worst_day_net": str(summary.worst_day_net),
-            "max_drawdown": str(drawdown),
-            "total_filled_notional": str(total_filled_notional),
-            "avg_filled_notional_per_day": str(avg_filled_notional_per_day),
-            "market_impact_liquidity_evidence_present": bool(daily_liquidity_notional),
-            "market_impact_liquidity_day_count": len(daily_liquidity_notional),
-            "market_impact_liquidity_missing_day_count": max(
-                0,
-                summary.trading_day_count - len(daily_liquidity_notional),
-            ),
-            "total_liquidity_notional": str(total_liquidity_notional),
-            "avg_liquidity_notional_per_day": str(avg_liquidity_notional_per_day),
-            "daily_net": {day: str(value) for day, value in summary.daily_net.items()},
-            "daily_filled_notional": {
-                day: str(value) for day, value in daily_filled_notional.items()
-            },
-            "daily_liquidity_notional": {
-                day: str(value) for day, value in daily_liquidity_notional.items()
-            },
-            "passed": not reasons,
-            "reasons": reasons,
-        },
-    )
-
-
-def _holdout_oos_passed(
-    *,
-    holdout_payload: Mapping[str, Any],
-    policy: ProfitabilityConstraintPolicy,
-) -> bool:
-    summary = summarize_replay_profitability(holdout_payload)
-    if policy.require_holdout_decisions and summary.decision_count <= 0:
-        return False
-    if summary.active_days < policy.min_active_holdout_days:
-        return False
-    if summary.net_per_day < policy.holdout_target_net_per_day:
-        return False
-    if summary.worst_day_net < -policy.max_worst_holdout_day_loss:
-        return False
-    if (
-        summary.profit_factor is None
-        or summary.profit_factor < policy.min_profit_factor
-    ):
-        return False
-    return True
 
 
 def _empty_replay_payload(*, start_date: date, end_date: date) -> dict[str, Any]:
